Escape quotes
We had an issue where a parent with a quote in their name (Eg. Robert "Bob" Jones) caused the Javascript string to end (line 130/131), which caused the "student" variable to not be set and the overall script to fail. Removing the quote before returning the value fixed it. This may need to be done on other fields as needed.